+++ CHANGELOG.md
@@ -1,5 +1,26 @@
 # Changelog
 
+## [7.3.1](https://github.com/supabase-community/realtime-csharp/compare/v7.3.0...v7.3.1) (2026-07-21)
+
+
+### Bug Fixes
+
+* resolve channel.Send() hanging on unacknowledged broadcast pushes ([#72](https://github.com/supabase-community/realtime-csharp/issues/72)) ([8f34dfe](https://github.com/supabase-community/realtime-csharp/commit/8f34dfee24594b1bc2c402a2649af12c3d1e0171))
+
+## [7.3.0](https://github.com/supabase-community/realtime-csharp/compare/v7.2.1...v7.3.0) (2026-07-20)
+
+
+### Features
+
+* attach Postgrest client context to models returned by PostgresChangesResponse ([#70](https://github.com/supabase-community/realtime-csharp/issues/70)) ([ce98568](https://github.com/supabase-community/realtime-csharp/commit/ce985685a328792c55d7f18d996469d2c860803e))
+
+## [7.2.1](https://github.com/supabase-community/realtime-csharp/compare/v7.2.0...v7.2.1) (2026-07-15)
+
+
+### Bug Fixes
+
+* lower Newtonsoft.Json minimum version to 13.0.2 ([#66](https://github.com/supabase-community/realtime-csharp/issues/66)) ([5d69eff](https://github.com/supabase-community/realtime-csharp/commit/5d69eff6fa54bee2e285f773e6a4c56935c772ba))
+
 ## [7.2.0](https://github.com/supabase-community/realtime-csharp/compare/v7.1.0...v7.2.0) (2025-05-13)